[swdb]: add transaction comparison test, unify TransactionItem objects
