+++ bindings/swig/swdb.i
@@ -9,7 +9,7 @@
 %include <std_shared_ptr.i>
 %include <std_string.i>
 %include <std_vector.i>
-%include <stdint.i>
+// %include <stdint.i>
 %include <typemaps.i>
 
 %shared_ptr(SQLite3)
@@ -39,6 +39,16 @@ typedef std::shared_ptr< libdnf::Transaction > TransactionPtr;
 %shared_ptr(TransactionItem)
 typedef std::shared_ptr< TransactionItem > TransactionItemPtr;
 
+%shared_ptr(MergedTransaction)
+typedef std::shared_ptr< MergedTransaction > MergedTransactionPtr;
+
+%shared_ptr(TransactionItemBase)
+typedef std::shared_ptr< TransactionItemBase > TransactionItemBasePtr;
+
+// FIXME this is a dirty workaround - file a patch to the swig upstream
+typedef long int int64_t;
+typedef unsigned int uint32_t;
+typedef int int32_t;
 
 %{
     // make SWIG wrap following headers
@@ -50,14 +60,19 @@ typedef std::shared_ptr< TransactionItem > TransactionItemPtr;
     #include "libdnf/swdb/swdb_types.hpp"
     #include "libdnf/swdb/transaction.hpp"
     #include "libdnf/swdb/transactionitem.hpp"
+    #include "libdnf/swdb/mergedtransaction.hpp"
 %}
 
 
 %template() std::vector<std::shared_ptr<libdnf::Transaction> >;
 %template() std::vector<std::shared_ptr<TransactionItem> >;
-
+%template() std::vector<std::shared_ptr<TransactionItemBase> >;
 %template() std::vector<std::shared_ptr<CompsGroupPackage> >;
 %template() std::vector<std::shared_ptr<CompsEnvironmentGroup> >;
+
+%template() std::vector<uint32_t>;
+%template() std::vector<int64_t>;
+%template() std::vector<bool>;
 
 %template() std::vector< std::string >;
 %template() std::pair<int,std::string>;
@@ -74,3 +89,4 @@ typedef std::shared_ptr< TransactionItem > TransactionItemPtr;
 %include "libdnf/swdb/swdb_types.hpp"
 %include "libdnf/swdb/transaction.hpp"
 %include "libdnf/swdb/transactionitem.hpp"
+%include "libdnf/swdb/mergedtransaction.hpp"
