+++ libdnf/swdb/private/CMakeLists.txt
@@ -1,5 +1,6 @@
 SET(SWDB_SRCS
     ${SWDB_SRCS}
     ${CMAKE_CURRENT_SOURCE_DIR}/repo.cpp
+    ${CMAKE_CURRENT_SOURCE_DIR}/transaction.cpp
     PARENT_SCOPE
 )
